remove duplicate DateTimeUtil
