+++ eagle-core/eagle-alert-parent/eagle-alert/alert-engine/src/main/java/org/apache/eagle/alert/engine/sorter/BaseStreamWindow.java
@@ -17,8 +17,9 @@
 package org.apache.eagle.alert.engine.sorter;
 
 import org.apache.eagle.alert.engine.PartitionedEventCollector;
-import org.apache.eagle.alert.utils.DateTimeUtil;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
+import org.apache.eagle.common.DateTimeUtil;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
